# contracts/cloud/openapi.yaml
openapi: 3.1.0
info:
  title: AO Cloud Public API
  version: 1.0.0
  description: |
    Client-facing contract for a separate AO Cloud implementation. This
    repository does not serve these routes. The worker surface documents the
    exact outbound HTTP protocol used by sandbox workers, including one-time
    bootstrap, rotating credentials, fenced turn execution, orchestration, and
    durable workspace and terminal transport.
security:
  - bearerAuth: []
...
paths:
...
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/terminal-ticket:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    post:
      operationId: createTerminalTicket
      tags: [Terminal]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind]
              properties:
                kind:
                  $ref: "#/components/schemas/TerminalKind"
      responses:
        "201":
          description: Short-lived, single-use terminal access ticket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalTicket"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/terminal:
    get:
      operationId: connectTerminal
      tags: [Terminal]
      summary: Upgrade to a terminal WebSocket using a short-lived ticket.
      security: []
      parameters:
        - name: ticket
          in: query
          required: true
          schema:
            type: string
        - name: after
          in: query
          schema:
            type: integer
            format: int64
            minimum: 0
            maximum: 9007199254740991
            default: 0
        - name: kind
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/TerminalKind"
      responses:
        "101":
          description: WebSocket protocol upgrade accepted.
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    workerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Session-scoped credential formatted as `Worker <token>`. Every
        authenticated route verifies both the token signature and that its
        worker ID and epoch are still current. Replacing a sandbox therefore
        fences an otherwise unexpired token with `STALE_WORKER_TOKEN`.

        Route scopes are `worker:connect` for heartbeat, `worker:event` for
        events, `worker:turn:claim`, `worker:turn:poll`, and
        `worker:turn:complete` for turn execution,
        `worker:credential:read` for the agent secret, `worker:git` for a
        checkout grant, `worker:orchestrate` for child controls, and
        `worker:transport` for workspace and terminal transport, and
        `worker:notification` for durable agent notification intake.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Standard Cloud API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, code, message, requestId]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
...
    TerminalKind:
      type: string
      enum: [agent, workspace]
    TerminalScope:
      type: string
      enum: [terminal:read, terminal:operate]
    TerminalTicket:
      type: object
      additionalProperties: false
      required: [ticket, expiresIn, scopes]
      properties:
        ticket:
          type: string
        expiresIn:
          type: integer
          minimum: 1
          description: Lifetime in seconds.
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/TerminalScope"
...